+++ migrations/m200114_220922_pages_parents.php
@@ -12,8 +12,10 @@ class m200114_220922_pages_parents extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn('{{%pages}}', 'parent_id', $this->integer(11)->null()->after('id'));
-        $this->createIndex('{{%idx-pages-parent}}', '{{%pages}}', ['parent_id']);
+        if (is_null($this->getDb()->getSchema()->getTableSchema('{{%pages}}')->getColumn('parent_id'))) {
+            $this->addColumn('{{%pages}}', 'parent_id', $this->integer(11)->null()->after('id'));
+            $this->createIndex('{{%idx-pages-parent}}', '{{%pages}}', ['parent_id']);
+        }
     }
 
     /**
@@ -21,7 +23,9 @@ class m200114_220922_pages_parents extends Migration
      */
     public function safeDown()
     {
-        $this->dropIndex('{{%idx-pages-parent}}', '{{%pages}}');
-        $this->dropColumn('{{%pages}}', 'parent_id');
+        if (!is_null($this->getDb()->getSchema()->getTableSchema('{{%pages}}')->getColumn('parent_id'))) {
+            $this->dropIndex('{{%idx-pages-parent}}', '{{%pages}}');
+            $this->dropColumn('{{%pages}}', 'parent_id');
+        }
     }
 }
